Move config types to a discrete file
